Fix CI and 32-bit builds
Also expand CI slightly and more targeted versions across Rust channels. Finally
also run `systest` on CI.
