+++ jemalloc-sys/Cargo.toml
@@ -7,10 +7,13 @@ links = "jemalloc"
 license = "MIT/Apache-2.0"
 repository = "https://github.com/alexcrichton/jemallocator"
 homepage = "https://github.com/alexcrichton/jemallocator"
-documentation = "http://alexcrichton.com/jemallocator"
+documentation = "https://docs.rs/jemallocator-sys"
 description = """
 Rust FFI bindings to jemalloc
 """
+
+[lib]
+test = false
 
 [dependencies]
 libc = { version = "0.2.8", default-features = false }
