+++ systest/Cargo.toml
@@ -9,4 +9,4 @@ jemalloc-sys = { path = "../jemalloc-sys" }
 libc = "0.1"
 
 [build-dependencies]
-ctest = { git = "https://github.com/alexcrichton/ctest" }
+ctest = "0.1"
