+++ systest/build.rs
@@ -8,6 +8,12 @@ fn main() {
 
     let mut cfg = ctest::TestGenerator::new();
     cfg.header("jemalloc/jemalloc.h")
-       .include(root.join("include"));
+       .include(root.join("include"))
+       .fn_cname(|rust, link_name| link_name.unwrap_or(rust).to_string());
+
+    if cfg!(target_os = "linux") {
+        cfg.skip_fn(|f| f == "malloc_usable_size");
+    }
+
     cfg.generate("../jemalloc-sys/src/lib.rs", "all.rs");
 }
